Wed, Dec 6, 2023, 9:48 PM -08:00

+++ wk2/pset2/readability/readability.c
@@ -31,7 +31,7 @@ int main(void)
     while (c != '\0');
     printf("s: %i w: %i l: %i\n", s, w, l);
     float L = (float)(l / w) * 100;
-    float S = (float)(s / w) * 100;
+    float S = (double)(s / w) * 100;
     float index = 0.0588 * L - 0.296 * S - 15.8;
     printf("L: %f, S: %f, Grade: %f\n", L, S, index);
 }
